Add search hint in search fields (close #137)
Under some circumstances there seems to be a bug in the behaviour
of the textfield. If it looses focus the emptyText string is displayed
but the textfield still has the input cursor. If I then click the
field again, the empty text is not deleted but treated as the normal text value
of the field. Let's wait and see if this bug gets annoying under normal usage scenarios.