Generic framework to cancel Ajax calls in front and backend.
The idea is to keep track of AJAX calls and the associated
PIDs of the backend process. An AJAX call can be canceled by
calling abort on the frontend. A function in the backend allows
to cancel (either by setting a flag or immediately kill) the associated
background process. AJAX call and backend process are
connected in a way that the AJAX call does not need information of
the PID and a process does not need knowledge of the AJAX call handle.

Currently implemented as proof of concept for the check_proxy function.