Fix subtle but super-nasty sha1 bug.
When Publication objects after a PDF match have only a doi set
the sha1 is undefined. After matching, the sha1 gets defined but
we restored to the undefined value in order to allow the updating
of the frontend when matching during PDF download occurs. But
when importing PDFs this means the undefined sha1 gets written into
the database. This currupts the database...

This commit should fix the issue and probably also the strange
duplicates search error.