Workaround for OS X Titanium cookie bug.
Cookies do not seem to be set/updated when requested
form a XMLHttp or even Iframe request. Only when a
Titanium window is pointed via the URL to the server the
cookie gets set. As a workaround we create an empty
Titanium window and load an empty page from the server
just to make sure that the cookie is set.