New backend for file sync.
We can't store the file sync information in the
Library database because if we sync between machines it is
unlikely that the user wants to sync with the same bibtex files on multiple
machines. So we store it in the User library which is meant to store settings
specific for a particular installation.