Set title when only booktitle is given in BibteX (fixes #563)
For Book and Inbook types a title must be set and not a
booktitle. Some BibTeX files get this wrong so we set the
title whenever it is missing in favour of the booktitle.

Seems only an issue for BibTeX at least I don't see any other
situation where somebody could set a Booktitle without title.