Try to fix race condition on startup.
I have added a message directly after the server is started
which we now use as control point when the server is ready.

We need to test whether this solves the issue. If so we
need to remember that we have to add this control line to
Catalyst/Engine/HTTP.pm whenever we update Catalyst.