Quote directory name in wrapper shell script.
It now works also in folders with spaces. Not so sure
about if LD_LIBRARY_PATH actually works also. In any case this is
kind of an edge case because no standard installation dir
has spaces. If somebody really has a custom installation dir with spaces
s/he is probably aware already that this might cause trouble under Linux.