Fix escaping issues with RSS feeds (closes #729).
The URL needs to be URI escaped before sent to the backend.
Otherwise this leads to unpredictable results.

I also had to make sure that the uiProvider field which holds
some JS  function is unset. Also that resulted in strange behaviour.
I'm not sure what uiProvide is exactly good for so I hope I did
not break anything here.