Fix many extremely annoying problems with Google Scholar import.
The fact that Google Scholar needs explicit completion of data cause
many problems. In particular, when it turns out that a paper is already
in the database after we have completed the data, we need to update
the guids in both the backend cache and the frontend. I hope this does
not get more complicated. It is reminiscent of the sha1 update hell
we had long time ago.